### 2018-11-13 #1080 @arfon
:wave: @telegraphic, thanks for your submission to JOSS. Our robot @whedon is somewhat fussy about filenames so could you please rename your `PAPER.md` to `paper.md` and update the paper to point to your bibliography (it's currently pointing to `paper.bib` but your bibtex appears to be in `references.bib`)
