### 2018-11-13 2c47459
Small changes to paper for JOSS submission
