+++ paper.md
@@ -13,7 +13,7 @@ authors:
   - name: Pieter T. Eendebak
     orcid: 0000-0001-7018-1124
     affiliation: "4, 5"
-  - Michael M. McKerns
+  - name: Michael M. McKerns
     orcid: 0000-0001-8342-3778
     affiliation: 6
   - name: Eben M. Olson
@@ -22,8 +22,6 @@ authors:
     affiliation: 8
   - name: Bairen Yi
     affiliation: 9
-
-    
 affiliations:
   - name: Department of Astronomy,  University of California Berkeley, Berkeley CA 94720
     index: 1
@@ -41,19 +39,18 @@ affiliations:
     index: 7
   - name: Google Brain, Mountain View, CA, 94043
     index: 8
-  - name: The Hong Kong University of Science and Technology 
+  - name: The Hong Kong University of Science and Technology
     index: 9
-
 date: 10 November 2018
 bibliography: paper.bib
 ---
 
-
+# Summary
 ``hickle`` is a Python 2/3 package for quickly dumping and loading python data structures to Hierarchical Data Format 5 (HDF5) files [@hdf5]. When dumping to HDF5, ``hickle`` automatically convert Python data structures (e.g. lists, dictionaries, ``numpy`` arrays [@numpy]) into HDF5 groups and datasets. When loading from file, ``hickle`` automatically converts data back into its original data type. A key motivation for ``hickle`` is to provide high-performance loading and storage of scientific data in the widely-supported HDF5 format.
 
-``hickle`` is designed as a drop-in replacement for the Python ``pickle`` package, which converts Python object hierarchies to and from Python-specific byte streams (processes known as 'pickling' and 'unpickling' respectively). Several different protocols exist, and files are not designed to be compatible between Python versions, nor interpretable in other languages. In contrast, ``hickle`` stores and loads files from HDF5, for which application programming interfaces (APIs) exist in most major languages, including C, Java, R, and MATLAB. 
+``hickle`` is designed as a drop-in replacement for the Python ``pickle`` package, which converts Python object hierarchies to and from Python-specific byte streams (processes known as 'pickling' and 'unpickling' respectively). Several different protocols exist, and files are not designed to be compatible between Python versions, nor interpretable in other languages. In contrast, ``hickle`` stores and loads files from HDF5, for which application programming interfaces (APIs) exist in most major languages, including C, Java, R, and MATLAB.
 
-Python data structures are mapped into the HDF5 abstract data model in a logical fashion, using the ``h5py`` package [@colette:2014]. Metadata required to reconstruct the hierarchy of objects, and to allow conversion into Python objects, is stored in HDF5 attributes. Most commonly used Python iterables (dict, tuple, list, set), and data types (int, float, str) are supported, as are ``numpy``  N-dimensional arrays. Commonly-used ``astropy`` data structures and ``scipy`` sparse matrices are also supported.
+Python data structures are mapped into the HDF5 abstract data model in a logical fashion, using the ``h5py`` package [@collette:2014]. Metadata required to reconstruct the hierarchy of objects, and to allow conversion into Python objects, is stored in HDF5 attributes. Most commonly used Python iterables (dict, tuple, list, set), and data types (int, float, str) are supported, as are ``numpy``  N-dimensional arrays. Commonly-used ``astropy`` data structures and ``scipy`` sparse matrices are also supported.
 
 ``hickle`` has been used in many scientific research projects, including:
 * Visualization and machine learning on volumetric fluorescence microscopy datasets from histological tissue imaging [@Durant:2017].
@@ -63,3 +60,5 @@ Python data structures are mapped into the HDF5 abstract data model in a logical
 
 ``hickle`` is released under the MIT license, and is available from PyPi via ``pip``; source code is available at
  https://github.com/telegraphic/hickle.
+
+# References
