### 2018-11-13 #1080 @arfon
@telegraphic - I made some small fixes to your paper here: https://github.com/telegraphic/hickle/pull/78
