### 2018-11-14 #1080 @telegraphic
Looks like the list at the end did not convert from markdown `*`s into bullets?
