### 2018-11-14 f11284e
Update paper.md (list formatting)
