### 2018-11-14 #1080 @telegraphic
An extra break seems to have done the trick, the proof looks great
