### 2018-11-14 #1080 @arfon
@telegraphic - could you take a look at [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few candidates that might be able to give this package a good review. Also, please start from somewhere near the bottom of the spreadsheet where some of the more recent JOSS review volunteers are listed.
