### 2018-11-14 #1080 @betteridiot
Let me give a quick glance. I will get back to you by 12:00 PM November 14th (lunch today).
