### 2018-12-03 #1080 @betteridiot
How is this going? Should I start reviewing?
