### 2018-12-04 #1115 @whedon
[REVIEW]: Hickle: A HDF5-based Python pickle replacement
**Submitting author:** @telegraphic (<a href="http://orcid.org/0000-0003-2783-1608">Danny Price</a>)
**Repository:** <a href="https://github.com/telegraphic/hickle" target ="_blank">https://github.com/telegraphic/hickle</a>
**Version:** v3.2.2
