### 2018-12-04 #1080 @arfon
@betteridiot @telegraphic see you both over in #1115 where the actual JOSS review takes place.
