### 2018-12-05 #1115 @betteridiot
@telegraphic, there is a DOI available for Raffel 2016: https://doi.org/10.7916/D8N58MHV
