### 2018-12-05 telegraphic/hickle#79 @betteridiot
JoSS Review: DOI for Raffel 2016
As mentioned in https://github.com/openjournals/joss-reviews/issues/1115, there is a DOI available for this paper:
https://doi.org/10.7916/D8N58MHV
