### 2018-12-05 telegraphic/hickle#80 @betteridiot
JoSS Review: No contributing guidelines
With respect to the last item of the *Documentation* checklist in https://github.com/openjournals/joss-reviews/issues/1115, a clear contribution guidelines should be furnished. A simple way of doing that can be found [here](https://help.github.com/articles/setting-guidelines-for-repository-contributors/#adding-a-contributing-file)
