### 2018-12-10 ca2a612
Create CONTRIBUTING.md contribution guidelines
As suggested in #80, adding some basic contribution guidelines
