### 2018-12-10 304cb5c
Update paper.bib
adding DOI for Raffel 2017
