+++ paper.bib
@@ -63,7 +63,8 @@
 	Author = {Colin Raffel},
 	School = {Columbia University},
 	Title = {Learning-Based Methods for Comparing Sequences, with Applications to Audio-to-MIDI Alignment and Matching},
-	Year = {2016}}
+	Year = {2016},
+	Doi = {https://doi.org/10.7916/D8N58MHV}}
 
 @inproceedings{Zhang:2016,
 	Acmid = {2934880},
