### 2018-12-10 #1115 @telegraphic
Thanks @betteridiot -- I've added contribution guidelines and the DOI for Raffel 2016
