### 2018-12-12 a58f59c
Python version compatibility
