### 2018-12-12 telegraphic/hickle#82 @betteridiot
JoSS Review: Documentation
Please see [this](https://github.com/betteridiot/joss_reviews/blob/master/hickle/joss_review.md) document for context.

This is more of a personal quibble, and @arfon should probably arbitrate this: while `hickle` is intended as a 'drop-in' replacement for `pickle`, it is actually an amalgam of `h5py` and `pickle/dill`. This means it has *extended* functionality. The issue I have with this is that the documentations kind of glosses over the `h5py` functionality with a link pointing to their documentation.

The reason this is an issue is that if `hickle` were to be a one stop shop for serializing to an HDF5 intermediate, I shouldn't have to go to someone else's documentation to do it. `hickle` is great! You should highlight this interoperability.

Lastly, with regards to documentation, I understand that `pickle` is very well-known. I also understand that (for those that use HDF5) `pytables` and `h5py` are well known as well. But there is no module level documentation for `hickle`. By that, I mean there is no reference to class or method documentation (docstrings).

I don't expect a [Read the Docs](https://readthedocs.org/) page for every package, but you already have the docstrings...a little time using [Sphinx autodoc](http://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html) can make your package very accessible.
