### 2018-12-12 telegraphic/hickle#83 @betteridiot
JoSS Review: Testing with `pytest`
I used `pytest` and the `pytest-coverage` plugin to run your automated tests.

Initially `pytest` could not run your test suite. I needed to install `astropy` and `scipy` to get it to run. These are not in plain Python or Miniconda environments by default. Since your package did not specify them as requirements, the tests could not run.

I recommend adding them to your requirements.txt
