### 2018-12-12 #1115 @betteridiot
Both https://github.com/telegraphic/hickle/issues/79 and https://github.com/telegraphic/hickle/issues/80 appear to be addressed by the author. These issues can be closed now.

I opened Issues https://github.com/telegraphic/hickle/issues/82 and https://github.com/telegraphic/hickle/issues/83 to address some specific items. I request your input @arfon on https://github.com/telegraphic/hickle/issues/82 specifically though.

I also made a PR at https://github.com/telegraphic/hickle/pull/81 to address some *very minor* compatibility issues.

Lastly, I would like to point @telegraphic  towards https://github.com/conda-forge/staged-recipes for a very simple way of adding `hickle` to conda-forge. This is what I did with [bamnostic](https://github.com/betteridiot/bamnostic), and it works great.

If you need any help or guidance, I am available. I believe that once these items are addressed, `hickle` should be good on my end.
