+++ hickle/loaders/load_pandas.py
@@ -2,4 +2,3 @@ import pandas as pd
 
 # TODO: populate with classes to load
 class_register = []
-exclude_register = []
