### 2018-12-13 d1760ee
Adding setup.py optional dependencies
Related to issue #83, making astropy/scipy optional dependencies.
Can now install e.g. hickle[astropy] to add astropy support.
Uses pkg_resources.requires('hickle[astropy') to check and only
load if error is not raised.
