### 2018-12-13 a85caf6
Changing from hickle import * line in __init__.py for tidier import
Now when using autocomplete in ipython or similar, far fewer methods
are visible to the end user (basically just load and dump, the core
functionality).
