+++ hickle/hickle.py
@@ -22,6 +22,7 @@ h5py installed.
 
 """
 
+from __future__ import absolute_import
 import sys
 import os
 from pkg_resources import get_distribution
@@ -59,6 +60,12 @@ except ModuleNotFoundError:
     import pickle
 
 import warnings
+
+from pkg_resources import get_distribution, DistributionNotFound
+try:
+    __version__ = get_distribution('hickle').version
+except DistributionNotFound:
+    __version__ = '0.0.0 - please install via pip/setup.py'
 
 ##################
 # Error handling #
