+++ README.md
@@ -1,34 +1,41 @@
 [![Build Status](https://travis-ci.org/telegraphic/hickle.svg?branch=master)](https://travis-ci.org/telegraphic/hickle)
+[![JOSS Status](http://joss.theoj.org/papers/0c6638f84a1a574913ed7c6dd1051847/status.svg)](http://joss.theoj.org/papers/0c6638f84a1a574913ed7c6dd1051847)
+
 
 Hickle
 ======
 
-Hickle is a HDF5 based clone of Pickle, with a twist. Instead of serializing to a pickle file,
-Hickle dumps to a HDF5 file. It is designed to be a "drop-in" replacement for pickle (for common data objects). 
-That is: it is a neat little way of dumping python variables to file. Hickle is fast, and allows for transparent compression of your data (LZF / GZIP).
+Hickle is a [HDF5](https://www.hdfgroup.org/solutions/hdf5/) based clone of `pickle`, with a twist: instead of serializing to a pickle file,
+Hickle dumps to a HDF5 file (Hierarchical Data Format). It is designed to be a "drop-in" replacement for pickle (for common data objects), but is
+really an amalgam of `h5py` and `dill`/`pickle` with extended functionality.
+
+That is: `hickle` is a neat little way of dumping python variables to HDF5 files that can be read in most programming 
+languages, not just Python. Hickle is fast, and allows for transparent compression of your data (LZF / GZIP). 
 
 Why use Hickle?
 ---------------
 
-While hickle is designed to be a drop-in replacement for pickle (and json), it works very differently. 
-Instead of serializing / json-izing, it instead stores the data using the excellent h5py module.
+While `hickle` is designed to be a drop-in replacement for `pickle` (or something like `json`), it works very differently. 
+Instead of serializing / json-izing, it instead stores the data using the excellent [h5py](https://www.h5py.org/) module.
 
 The main reasons to use hickle are:
 
-  1. it's faster than pickle and cPickle
-  2. it stores data in HDF5
+  1. It's faster than pickle and cPickle.
+  2. It stores data in HDF5.
   3. You can easily compress your data.
 
 The main reasons not to use hickle are:
 
   1. You don't want to store your data in HDF5. While hickle can serialize arbitrary python objects, this functionality is provided only for convenience, and you're probably better off just using the pickle module.
-  2. You want to convert your data in JSON. For this, use a json or uJson.
+  2. You want to convert your data in human-readable JSON/YAML, in which case, you should do that instead.
 
-So, if you want your data in HDF5, or if your pickling is taking too long, give hickle a try. Hickle is particularly good at storing large numpy arrays, thanks to h5py running under the hood. 
+So, if you want your data in HDF5, or if your pickling is taking too long, give hickle a try. 
+Hickle is particularly good at storing large numpy arrays, thanks to `h5py` running under the hood. 
 
 Recent changes
 --------------
 
+* November 2018: Submitted to Journal of Open-Source Software (JOSS).
 * June 2018: Major refactor and support for Python 3. 
 * Aug 2016: Added support for scipy sparse matrices `bsr_matrix`, `csr_matrix` and `csc_matrix`.
 
@@ -79,7 +86,7 @@ For storing python dictionaries of lists, hickle beats the python json encoder, 
 It should be noted that these comparisons are of course not fair: storing in HDF5 will not help you convert something into JSON, nor will it help you serialize a string. But for quick storage of the contents of a python variable, it's a pretty good option.
 
 Installation guidelines (for Linux and Mac OS).
-----------------------------------------------------------------------------------------------------
+-----------------------------------------------
 
 ### Easy method
[…]
