### 2018-12-13 telegraphic/hickle#83 @telegraphic
Hi @betteridiot, this was the push I needed to figure out the ' square bracket install thing' (which I now know is called an 'extra optional dependency'). Not everyone will need/want astropy. Do you know if there's a good way to support this in unit testing?
