### 2018-12-13 telegraphic/hickle#82 @telegraphic
Hi @betteridiot, I've updated the README.md with more information to address some of your concerns. I don't disagree that a spiffy readthedocs page would be excellent, but at the present moment I am worried about the extra time required to create and maintain it... So, my preference would be to decouple this from the JOSS review, and hold off on this until I (or another contributor #pull-request-welcome!) have more free time.
