### 2018-12-13 #1115 @telegraphic
Thanks @betteridiot, I very much appreciate your time on this! I've tried to address all the issues and improved the documentation in the README.md. Working on the conda recipe now
