+++ setup.py
@@ -5,12 +5,16 @@
 # python setup.py sdist upload
 from setuptools import setup, find_packages
 
-version = '3.3.0'
+version = '3.3.1'
 author  = 'Danny Price'
+
+with open("README.md", "r") as fh:
+    long_description = fh.read()
 
 setup(name='hickle',
       version=version,
       description='Hickle - a HDF5 based version of pickle',
+      long_description=long_description,
       author=author,
       author_email='dan@thetelegraphic.com',
       url='http://github.com/telegraphic/hickle',
