### 2018-12-13 0d545dc
Changes for pylint checking
