### 2018-12-13 bf41835
Added `tests_require`
This new setup will allow the end-user to do `python setup.py test` to run `pytest` on the test suite
