### 2018-12-13 1e7ac31
Add aliasing and verbosity setting to pytest
This controls how `setuptools` works when `python setup.py test` is invoked. By aliasing `test` to `pytest`, it will run `pytest` instead of `unittest` by default. This doesn't change implementation since `pytest` is compatible with `unittest`. Also, it add a verbosity tag by default when `python setup.py test` is invoked.
