+++ setup.cfg
@@ -1,2 +1,8 @@
 [metadata]
-description-file = README.md
+description-file=README.md
+
+[aliases]
+test=pytest
+
+[tool:pytest]
+addopts=--verbose
