### 2018-12-13 d95f2c0
Added automated coverage testing
