+++ setup.cfg
@@ -5,4 +5,4 @@ description-file=README.md
 test=pytest
 
 [tool:pytest]
-addopts=--verbose
+addopts=--verbose --cov
