+++ setup.py
@@ -19,7 +19,7 @@ setup(name='hickle',
       keywords=['pickle', 'hdf5', 'data storage', 'data export'],
       #py_modules = ['hickle', 'hickle_legacy'],
       install_requires=['numpy', 'h5py'],
-      setup_requires = ['pytest-runner',],
+      setup_requires = ['pytest-runner', 'pytest-cov'],
       tests_require = ['pytest', 'astropy', 'scipy', 'pandas'],
       python_requires='>=2.7',
       packages=find_packages(),
