+++ .travis.yml
@@ -17,7 +17,7 @@ install:
   - sudo apt-get update -qq
   - sudo apt-get install -qq libhdf5-serial-dev
   - pip install pytest coveralls
-  - pip install --only-binary=numpy
+  - pip install --only-binary=numpy numpy
   - pip install -r requirements.txt
   - pip install .
 
