+++ setup.py
@@ -20,7 +20,7 @@ setup(name='hickle',
       #py_modules = ['hickle', 'hickle_legacy'],
       install_requires=['numpy', 'h5py'],
       setup_requires = ['pytest-runner', 'pytest-cov'],
-      tests_require = ['pytest', 'astropy', 'scipy', 'pandas'],
+      tests_require = ['pytest', 'astropy < 3.0 ; python_version < "3.0"', 'astropy == 3.0 ; "2.7" < python_version < "3.5"', 'astropy >= 3.1 ; python_version >= "3.5"', 'scipy', 'pandas'],
       python_requires='>=2.7',
       packages=find_packages(),
       zip_safe=False,
