### 2018-12-13 ce51c55
Update setup.py
