+++ .travis.yml
@@ -16,7 +16,7 @@ matrix:
 install:
   - sudo apt-get update -qq
   - sudo apt-get install -qq libhdf5-serial-dev
-  - pip install pytest coveralls
+  - pip install pytest coveralls setuptools<=36.2
   - pip install --only-binary=numpy numpy
   - pip install -r requirements.txt
   - pip install .
