### 2018-12-13 157602c
Update setup.py
