### 2018-12-13 1e670a3
Update setup.py
