+++ setup.py
@@ -8,7 +8,7 @@ import sys
 
 if sys.version_info.major == 3:
       if sys.version_info.minor < 5:
-            astro = "astropy<=3.0
+            astro = "astropy<=3.0"
       else:
             astro = "astropy>=3.2"
 else:
