+++ html/_sources/index.rst.txt
@@ -1,20 +1,0 @@
-.. hickle documentation master file, created by
-   sphinx-quickstart on Fri Dec 14 15:39:45 2018.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to hickle's documentation!
-==================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
