+++ docs/source/index.rst
@@ -10,6 +10,8 @@ Welcome to hickle's documentation!
    :maxdepth: 2
    :caption: Contents:
 
+.. automodule::
+   :members: load, dump
 
 
 Indices and tables
