### 2018-12-14 db9ce11
Adding source setup sphinx
