### 2018-12-14 #1115 @telegraphic
conda recipe is working! https://github.com/conda-forge/staged-recipes/pull/7268
