+++ generate_docs.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+git checkout gh-pages
+cp ../hickle/README.md ./
+cp -r ../hickle/docs/build/html/* ./
+
+git add .
+git commit -m 'Update docs `date`'
+
