+++ generate_docs.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
+cd ../hickle/docs
+make html
+cd ../../hickle_docs
+
 git checkout gh-pages
 cp ../hickle/README.md ./
 cp -r ../hickle/docs/build/html/* ./
