+++ _sources/index.md.txt
@@ -1,11 +1,9 @@
-.. hickle documentation master file, created by
-   sphinx-quickstart on Fri Dec 14 15:39:45 2018.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
+[![Build Status](https://travis-ci.org/telegraphic/hickle.svg?branch=master)](https://travis-ci.org/telegraphic/hickle)
+[![JOSS Status](http://joss.theoj.org/papers/0c6638f84a1a574913ed7c6dd1051847/status.svg)](http://joss.theoj.org/papers/0c6638f84a1a574913ed7c6dd1051847)
 
-Welcome to hickle's documentation!
-==================================
 
+Hickle
+======
 
 Hickle is a [HDF5](https://www.hdfgroup.org/solutions/hdf5/) based clone of `pickle`, with a twist: instead of serializing to a pickle file,
 Hickle dumps to a HDF5 file (Hierarchical Data Format). It is designed to be a "drop-in" replacement for pickle (for common data objects), but is
@@ -14,19 +12,202 @@ really an amalgam of `h5py` and `dill`/`pickle` with extended functionality.
 That is: `hickle` is a neat little way of dumping python variables to HDF5 files that can be read in most programming 
 languages, not just Python. Hickle is fast, and allows for transparent compression of your data (LZF / GZIP). 
 
+Why use Hickle?
+---------------
+
+While `hickle` is designed to be a drop-in replacement for `pickle` (or something like `json`), it works very differently. 
+Instead of serializing / json-izing, it instead stores the data using the excellent [h5py](https://www.h5py.org/) module.
+
+The main reasons to use hickle are:
+
+  1. It's faster than pickle and cPickle.
+  2. It stores data in HDF5.
+  3. You can easily compress your data.
+
+The main reasons not to use hickle are:
+
+  1. You don't want to store your data in HDF5. While hickle can serialize arbitrary python objects, this functionality is provided only for convenience, and you're probably better off just using the pickle module.
+  2. You want to convert your data in human-readable JSON/YAML, in which case, you should do that instead.
+
+So, if you want your data in HDF5, or if your pickling is taking too long, give hickle a try. 
+Hickle is particularly good at storing large numpy arrays, thanks to `h5py` running under the hood. 
+
+Recent changes
+--------------
+
+* November 2018: Submitted to Journal of Open-Source Software (JOSS).
+* June 2018: Major refactor and support for Python 3. 
+* Aug 2016: Added support for scipy sparse matrices `bsr_matrix`, `csr_matrix` and `csc_matrix`.
+
+Performance comparison
+----------------------
+
+Hickle runs a lot faster than pickle with its default settings, and a little faster than pickle with `protocol=2` set:
+
+```Python
+In [1]: import numpy as np
+
+In [2]: x = np.random.random((2000, 2000))
+
+In [3]: import pickle
+
+In [4]: f = open('foo.pkl', 'w')
+
+In [5]: %time pickle.dump(x, f)  # slow by default
+CPU times: user 2 s, sys: 274 ms, total: 2.27 s
+Wall time: 2.74 s
+
+In [6]: f = open('foo.pkl', 'w')
+
+In [7]: %time pickle.dump(x, f, protocol=2)  # actually very fast
+CPU times: user 18.8 ms, sys: 36 ms, total: 54.8 ms
+Wall time: 55.6 ms
+
+In [8]: import hickle
+
+In [9]: f = open('foo.hkl', 'w')
+
+In [10]: %time hickle.dump(x, f)  # a bit faster
+dumping <type 'numpy.ndarray'> to file <HDF5 file "foo.hkl" (mode r+)>
+CPU times: user 764 µs, sys: 35.6 ms, total: 36.4 ms
+Wall time: 36.2 ms
+```
+
[…]
