### 2018-12-14 telegraphic/hickle#83 @betteridiot
Now that the tests are up and going. You should probably add a "Testing" section to your README.md to prompt users to run `python setup.py test` when installed from source.
