### 2018-12-14 #1115 @arfon
> @arfon Looks good on my end.

Great, thanks @betteridiot.

> I opened Issues telegraphic/hickle#82 and telegraphic/hickle#83 to address some specific items. I request your input @arfon on telegraphic/hickle#82 specifically though.

I think where we've got to here is good with me.
