### 2018-12-17 #1115 @telegraphic
Hi @arfon -- record is https://zenodo.org/record/2345649, and DOI is 10.5281/zenodo.2345649.

@betteridiot -- after some fiddling I've gotten a conda-forge recipe too, waiting on checks (https://github.com/conda-forge/staged-recipes/pull/7268)
