+++ _sources/toc.rst.txt
@@ -7,7 +7,7 @@ Welcome to hickle's documentation!
 ==================================
 
 
-Hickle is a [HDF5](https://www.hdfgroup.org/solutions/hdf5/) based clone of `pickle`, with a twist: instead of serializing to a pickle file,
+Hickle is a HDF5-based clone of `pickle`, with a twist: instead of serializing to a pickle file,
 Hickle dumps to a HDF5 file (Hierarchical Data Format). It is designed to be a "drop-in" replacement for pickle (for common data objects), but is
 really an amalgam of `h5py` and `dill`/`pickle` with extended functionality.
 
