### 2018-12-17 50d425d
Updated documentation
