+++ docs/make_docs.sh
@@ -1,0 +1,2 @@
+cp ../README.md ./source/index.md
+make html
