+++ docs/source/index.rst
@@ -1,22 +1,0 @@
-.. hickle documentation master file, created by
-   sphinx-quickstart on Fri Dec 14 15:39:45 2018.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to hickle's documentation!
-==================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-.. automodule:: hickle
-   :members: load, dump
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
