### 2018-12-17 telegraphic/hickle#83 @telegraphic
Alrighty, I've added this to README.md
